requirements.txt
bleach==3.3.0
build==0.3.1.post1
certifi==2020.12.5
chardet==4.0.0
cognitojwt==1.2.2
colorama==0.4.4
docutils==0.16
ecdsa==0.14.1
fastapi==0.63.0
idna==2.10
importlib-metadata==3.10.0
keyring==23.0.1
packaging==20.9
pep517==0.10.0
pkginfo==1.7.0
pyasn1==0.4.8
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-jose==3.2.0
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
starlette==0.13.6
toml==0.10.2
tqdm==4.59.0
twine==3.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
zipp==3.4.1
starlette-context==0.3.1
